requirements.txt
bingX_connector==0.0.7
firebase_admin==6.2.0
gate_api==4.60.2
kucoin_futures_python==1.0.8
matplotlib==3.8.0
mplfinance==0.12.10b0
numpy==1.26.3
pandas==2.1.4
protobuf==4.25.1
python-decouple==3.8
python-telegram-bot==20.7
python_bitget==1.0.7
Requests==2.31.0
TA_Lib==0.4.28
